sofdesk/pyproject.toml
[tool.poetry]
name = "sofdesk"
version = "v1"
description = ""
authors = ["lgd"]
readme = ""

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
django-debug-toolbar = "^4.1.0"
drf-nested-routers = "^0.93.4"
django-dry-rest-permissions = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
